+++ src/main/php/xp/compiler/io/FileSource.class.php
@@ -2,10 +2,10 @@
 
 use io\File;
 use xp\compiler\Syntax;
+use lang\IllegalArgumentException;
 
 /**
  * Source implementation based on files
- *
  */
 class FileSource extends \lang\Object implements Source {
   protected $file= null;
@@ -16,10 +16,15 @@ class FileSource extends \lang\Object implements Source {
    *
    * @param   io.File file
    * @param   xp.compiler.Syntax s Syntax to use, determined via source file's syntax otherwise
+   * @throws  lang.IllegalArgumentException in case the syntax cannot be determined
    */
   public function __construct(File $file, Syntax $s= null) {
     $this->file= $file;
-    $this->syntax= $s ?: Syntax::forName($this->file->getExtension());
+    try {
+      $this->syntax= $s ?: Syntax::forName($this->file->getExtension());
+    } catch (IllegalArgumentException $e) {
+      throw new IllegalArgumentException('Cannot determine syntax for "'.$this->file->getFileName().'"', $e);
+    }
   }
   
   /**